feat(rog-control-center): move startup to helper and populate asus_armoury_loaded property
