+++ rog-control-center/src/helpers/mod.rs
@@ -1,3 +1,4 @@
 pub mod hardware;
+pub mod startup;
 pub mod types;
 pub mod zbus_proxies;
