+++ rog-control-center/src/lib.rs
@@ -10,7 +10,6 @@ pub mod cli_options;
 pub mod config;
 pub mod error;
 pub mod helpers;
-pub mod startup;
 pub mod state;
 pub mod tray;
 pub mod ui;
